Moved django.vim into ftplugin/, and replaced with the original from vim scripts.
The previous ./django.vim was the same as the original except for the comments, and being wrapped in an `if`. Since Vim's ftplugin/htmldjango.vim sources ftplugin/django.vim (see https://code.google.com/p/vim/source/browse/runtime/ftplugin/htmldjango.vim) all that was needed was to move it to the right place.